No longer color whitespace, cleaned up HtmlDecorator, moved formatMetrics to ShardTableQueryMetricHandler, deleted JexlFormattedStringBuildingVisitor and JexlFormattedStringBuildingVisitorTest: moved all functionality to JexlStringBuildingVisitor and JexlStringBuildingVisitorTest
