+++ warehouse/query-core/src/main/java/datawave/query/metrics/ShardTableQueryMetricHandler.java
@@ -82,7 +82,11 @@ import datawave.microservice.querymetric.QueryMetricListResponse;
 import datawave.microservice.querymetric.QueryMetricsDetailListResponse;
 import datawave.microservice.querymetric.QueryMetricsSummaryResponse;
 import datawave.query.iterator.QueryOptions;
-import datawave.query.jexl.visitors.JexlFormattedStringBuildingVisitor;
+import datawave.query.jexl.JexlASTHelper;
+import datawave.query.jexl.visitors.HtmlDecorator;
+import datawave.query.jexl.visitors.JexlQueryDecorator;
+import datawave.query.jexl.visitors.JexlStringBuildingVisitor;
+import datawave.query.jexl.visitors.PrintingVisitor;
 import datawave.query.language.parser.jexl.LuceneToJexlQueryParser;
 import datawave.query.map.SimpleQueryGeometryHandler;
 import datawave.security.authorization.DatawavePrincipal;
@@ -551,7 +555,7 @@ public class ShardTableQueryMetricHandler extends BaseQueryMetricHandler<QueryMe
             query.setParameters(ImmutableMap.of(QueryOptions.INCLUDE_GROUPING_CONTEXT, "true"));
             List<QueryMetric> queryMetrics = getQueryMetrics(response, query, datawavePrincipal);
             
-            JexlFormattedStringBuildingVisitor.formatMetrics(queryMetrics);
+            formatMetrics(queryMetrics);
             
             response.setResult(queryMetrics);
             
@@ -562,8 +566,55 @@ public class ShardTableQueryMetricHandler extends BaseQueryMetricHandler<QueryMe
 
         return response;
     }
-
-    public BaseQueryMetric toMetric(EventBase event) {
+    
+    /**
+     * Returns a new list of QueryMetrics which is identical to the given list except the query and query plan for each metric are formatted according to
+     * {@link datawave.query.jexl.visitors.JexlStringBuildingVisitor#buildDecoratedQuery buildDecoratedQuery} with an
+     * {@link datawave.query.jexl.visitors.HtmlDecorator HtmlDecorator}
+     * 
+     * @param metrics
+     */
+    public static void formatMetrics(List<? extends BaseQueryMetric> metrics) {
+        // For each metric, update the query to be formatted (if applicable) and update
+        // the plan to be formatted
+        for (BaseQueryMetric metric : metrics) {
+            JexlNode queryNode = null, planNode = null;
+            String query = metric.getQuery();
+            String plan = metric.getPlan();
+            // If it is a JEXL query, set the query to be formatted
+            if (query != null && isJexlQuery(metric.getParameters())) {
+                try {
+                    queryNode = JexlASTHelper.parseJexlQuery(query);
+                    metric.setQuery(JexlStringBuildingVisitor.buildDecoratedQuery(queryNode, false, new HtmlDecorator()));
+                } catch (org.apache.commons.jexl2.parser.ParseException e) {
+                    log.error("Could not parse JEXL AST after performing transformations to run the query", e);
+                    
+                    if (log.isTraceEnabled()) {
+                        log.trace(PrintingVisitor.formattedQueryString(queryNode));
+                    }
+                }
+            }
+            // Format the plan (plan will always be a JEXL query)
+            if (plan != null) {
+                try {
+                    planNode = JexlASTHelper.parseJexlQuery(plan);
+                    metric.setPlan(JexlStringBuildingVisitor.buildDecoratedQuery(planNode, false, new HtmlDecorator()));
+                } catch (org.apache.commons.jexl2.parser.ParseException e) {
+                    log.error("Could not parse JEXL AST after performing transformations to run the query", e);
+                    
+                    if (log.isTraceEnabled()) {
+                        log.trace(PrintingVisitor.formattedQueryString(planNode));
+                    }
+                }
+            }
+        }
+    }
+    
+    private static boolean isJexlQuery(Set<Parameter> params) {
+        return params.stream().anyMatch(p -> p.getParameterName().equals("query.syntax") && p.getParameterValue().equals("JEXL"));
+    }
+    
+    public QueryMetric toMetric(datawave.webservice.query.result.event.EventBase event) {
         SimpleDateFormat sdf_date_time1 = new SimpleDateFormat("yyyyMMdd HHmmss");
         SimpleDateFormat sdf_date_time2 = new SimpleDateFormat("yyyyMMdd HHmmss");
         SimpleDateFormat sdf_date_time3 = new SimpleDateFormat("yyyyMMdd");
